// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };

/**
 * Item in *News Post → sections*
 */
export interface NewsPostDocumentDataSectionsItem {
  /**
   * title field in *News Post → sections*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: news_post.sections[].title
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  title: prismic.KeyTextField;

  /**
   * descriptions field in *News Post → sections*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: news_post.sections[].descriptions
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  descriptions: prismic.RichTextField;
}

type NewsPostDocumentDataSlicesSlice = never;

/**
 * Content for News Post documents
 */
interface NewsPostDocumentData {
  /**
   * title field in *News Post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Title
   * - **API ID Path**: news_post.title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  title: prismic.KeyTextField;

  /**
   * date field in *News Post*
   *
   * - **Field Type**: Date
   * - **Placeholder**: *None*
   * - **API ID Path**: news_post.date
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#date
   */
  date: prismic.DateField;

  /**
   * Image field in *News Post*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: news_post.image
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  image: prismic.ImageField<never>;

  /**
   * sections field in *News Post*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: news_post.sections[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  sections: prismic.GroupField<Simplify<NewsPostDocumentDataSectionsItem>>;

  /**
   * Slice Zone field in *News Post*
   *
   * - **Field Type**: Slice Zone
   * - **Placeholder**: *None*
   * - **API ID Path**: news_post.slices[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#slices
   */
  slices: prismic.SliceZone<NewsPostDocumentDataSlicesSlice> /**
   * Meta Title field in *News Post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A title of the page used for social media and search engines
   * - **API ID Path**: news_post.meta_title
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */;
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *News Post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A brief summary of the page
   * - **API ID Path**: news_post.meta_description
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  meta_description: prismic.KeyTextField;

  /**
   * Meta Image field in *News Post*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: news_post.meta_image
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  meta_image: prismic.ImageField<never>;
}

/**
 * News Post document from Prismic
 *
 * - **API ID**: `news_post`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type NewsPostDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<NewsPostDocumentData>,
    "news_post",
    Lang
  >;